pyproject.toml
[project]
name = "company-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "typer>=0.16.0",
]
